refactor : moving controllers and middlewares to the presentation layer of the project
