+++ src/errors/errors.ts
@@ -1,9 +1,19 @@
 export class AppError extends Error {
-  code: number;
-  constructor(message: string, code: number) {
+  statusCode: number;
+  constructor(message: string, statusCode: number) {
     super();
     this.message = message;
-    this.code = code;
+    this.statusCode = statusCode;
     this.name = "AppError";
+  }
+}
+
+export class ResourseAlreadyExistError extends Error{
+  statusCode: number;
+  constructor(message: string, statusCode: number) {
+    super();
+    this.message = message;
+    this.statusCode = statusCode;
+    this.name = "ResourseAlreadyExistError";
   }
 }
